+++ MOP/src/Places/Cases/Yard.cs
@@ -49,12 +49,18 @@ namespace MOP.Places
             "WC", "Hallway", "Entry", "ContactPivot", "DoorRight", "DoorLeft", "GarageDoors", "BatteryCharger",
             "Clamps", "ChargerPivot", "Clamp", "BatteryPivot", "battery_charger", "Wire", "cable", "TriggerCharger",
             "tvtable", "VHS_Screen", "tv_table(Clone)", "scart_con", "Haybale", "Combine", "UncleWalking", "LOD", "house_wall_brick",
-            "houseuncle_roof", "houseuncle_walls", "fuse holder(Clone)"
+            "houseuncle_roof", "houseuncle_walls", "fuse holder(Clone)", "SAUNA"
         };
 
         const float ChillDistance = .45f;
         readonly Transform chillPoint;
         readonly FsmBool fridgeRunning;
+        
+        GameObject sauna;
+        GameObject saumaSimulation;
+        readonly FsmFloat saunaStoveHeat;
+        const float StoveOnSimulationPoint = 35; // Stove heat after which we will simulate stove overheating
+
 
         /// <summary>
         /// Initialize the RepairShop class
@@ -65,12 +71,8 @@ namespace MOP.Places
 
             // Fix for broken computer.
             // We're changing it to null.
-            if (GameObject.Find("COMPUTER") != null)
-                GameObject.Find("COMPUTER").transform.parent = null;
-
-            GameObject garageDoors = GameObject.Find("GarageDoors");
-            garageDoors.transform.Find("DoorLeft/Coll").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
-            garageDoors.transform.Find("DoorRight/Coll").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
+            RemoveComputerFromHome();
+            FixGarageDoors();
 
             try
             {
@@ -94,23 +96,64 @@ namespace MOP.Places
 
             // Disable restarting of FSM in UNCLE object.
             transform.Find("UNCLE").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
+
             Compress();
 
-            transform.Find("Building/SAUNA/Sauna/Kiuas/ButtonPower").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
-            transform.Find("Building/SAUNA/Sauna/Kiuas/ButtonTime").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
-            transform.Find("Building/SAUNA/Sauna/Kiuas/StoveTrigger").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
-            transform.Find("Building/SAUNA/Sauna/Simulation").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
+            FixSaunaSimulation();
 
             chillPoint = transform.Find("Building/KITCHEN/Fridge/FridgePoint/ChillArea");
             try
             {
-                fridgeRunning = Resources.FindObjectsOfTypeAll<GameObject>()?.First(g => g.name == "FridgePoint").GetPlayMaker("Chilling")?.FsmVariables.GetFsmBool("Kitchen");
+                GameObject fridgePoint = Resources.FindObjectsOfTypeAll<GameObject>().FirstOrDefault(g => g.name == "FridgePoint");
+                if (fridgePoint)
+                {
+                    fridgeRunning = fridgePoint.GetPlayMaker("Chilling")?.FsmVariables.GetFsmBool("Kitchen");
+                }
             }
             catch (Exception ex)
             {
                 ExceptionManager.New(ex, false, "FRIDGE_RUNNING_FAILURE");
             }
+
+            // Get sauna simulation.
+            try
+            {
+                sauna = transform.Find("Building/SAUNA")?.gameObject;
+                if (sauna != null)
+                {
+                    saumaSimulation = sauna.transform.Find("Sauna/Simulation")?.gameObject;
+                    if (saumaSimulation != null)
+                    {
+                        saunaStoveHeat = saumaSimulation.GetPlayMaker("Time").FsmVariables.GetFsmFloat("StoveHeat");
+                    }
+                }
             }
+            catch (Exception ex)
+            {
+                ExceptionManager.New(ex, false, "SAUNA_STOVE_SIMULATION_FAILURE");
+            }
+
+            LightSources = GetLightSources();
+        }
+
+        private static void FixGarageDoors()
+        {
+            GameObject garageDoors = GameObject.Find("GarageDoors");
+            if (garageDoors)
+            {
+                garageDoors.transform.Find("DoorLeft/Coll").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
+                garageDoors.transform.Find("DoorRight/Coll").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
+            }
+        }
+
+        private static void RemoveComputerFromHome()
+        {
+            GameObject computer = GameObject.Find("COMPUTER");
+            if (computer != null)
+            {
+                computer.transform.parent = null;
+            }
+        }
 
         public bool IsItemInFridge(GameObject item)
         {
@@ -121,6 +164,74 @@ namespace MOP.Places
                 return false;
 
             return Vector3.Distance(item.transform.position, chillPoint.position) < ChillDistance;
+        }
+
+        private void FixSaunaSimulation()
+        {
+            if (transform.Find("Building/SAUNA") != null)
+            {
+                transform.Find("Building/SAUNA/Sauna/Kiuas/ButtonPower").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
+                transform.Find("Building/SAUNA/Sauna/Kiuas/ButtonTime").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
+                transform.Find("Building/SAUNA/Sauna/Kiuas/StoveTrigger").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
+                transform.Find("Building/SAUNA/Sauna/Simulation").GetComponent<PlayMakerFSM>().Fsm.RestartOnEnable = false;
+            }
+        }
+
+        public override void ToggleActive(bool enabled)
+        {
+            // Don't execute the code, if the enabled value is the same as the activity status.
+            if (isActive == enabled)
+                return;
+
+            isActive = enabled;
+
+            // Load and unload only the objects that aren't on the whitelist.
+            for (int i = 0; i < DisableableChilds.Count; i++)
+            {
+                // If the object is missing, skip and continue.
+                if (DisableableChilds[i] == null)
+                    continue;
+
+                DisableableChilds[i].gameObject.SetActive(enabled);
+            }
+
+            if (PlayMakers.Count > 0)
+            {
+                for (int i = 0; i < PlayMakers.Count; i++)
+                {
+                    if (PlayMakers[i] == null)
+                    {
+                        continue;
+                    }
+
+                    PlayMakers[i].enabled = enabled;
+                }
+            }
+
+            if (LightSources.Count > 0)
+            {
+                if (FsmManager.ShadowsHouse)
+                {
+                    for (int i = 0; i < LightSources.Count; ++i)
+                    {
+                        LightSources[i].shadows = enabled ? LightShadows.Hard : LightShadows.None;
+                    }
+                }
+            }
+
+            if (sauna != null)
+            {
+                if (saunaStoveHeat.Value > StoveOnSimulationPoint)
+                {
+                    sauna.SetActive(true);
+                    saumaSimulation.SetActive(true);
+                }
+                else
+                {
+                    sauna.SetActive(enabled);
+                    saumaSimulation.SetActive(enabled);
+                }
+            }
         }
     }
 }