+++ MOP/src/Rules/RulesManager.cs
@@ -114,9 +114,9 @@ namespace MOP.Rules
             foreach (Rule rule in Rules)
             {
 #if PRO
-                if (!rule.Mod.Enabled) continue;
+                if (rule.Mod != null && !rule.Mod.Enabled) continue;
 #else
-                if (rule.Mod.isDisabled) continue;
+                if (rule.Mod != null && rule.Mod.isDisabled) continue;
 #endif
 
                 if (rule.GetType() == typeof(T))