+++ MOP/src/WorldObjects/SimpleObjectToggle.cs
@@ -23,14 +23,16 @@ namespace MOP.WorldObjects
 {
     class SimpleObjectToggle : GenericObject
     {
-        public SimpleObjectToggle(GameObject gameObject, DisableOn disableOn = DisableOn.Distance, int distance = 200) : base(gameObject, distance, disableOn)
-        {
-            return;
-        }
+        public SimpleObjectToggle(GameObject gameObject, DisableOn disableOn = DisableOn.Distance, int distance = 200) : base(gameObject, distance, disableOn) { }
 
         public override void Toggle(bool enabled)
         {
             if (DisableOn.HasFlag(DisableOn.IgnoreInQualityMode) && MopSettings.Mode == PerformanceMode.Quality)
+            {
+                enabled = true;
+            }
+
+            if (DisableOn.HasFlag(DisableOn.IgnoreInBalancedAndAbove) && MopSettings.Mode >= PerformanceMode.Balanced)
             {
                 enabled = true;
             }